+++ src/util/polyfills.js
@@ -3,36 +3,9 @@
 //
 
 // Smoothscroll
+// TODO: scroll-behaviour smooth was added to Safari on March 2022.
+//       We could consider removing this on 2023.
 require('smoothscroll-polyfill').polyfill();
-
-// [].includes
-require('array-includes').shim();
-
-// [].find
-require('array.prototype.find').shim();
-
-// Object.entries
-require('object.entries').shim();
-
-// Object.values
-require('object.values').shim();
-
-// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/parseFloat
-if (typeof Number.parseFloat === 'undefined' && typeof window !== 'undefined') {
-  Number.parseFloat = window.parseFloat;
-}
-
-// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/parseInt
-if (typeof Number.parseInt === 'undefined' && typeof window !== 'undefined') {
-  Number.parseInt = window.parseInt;
-}
-
-// NaN is the only value in javascript which is not equal to itself.
-// https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/isNaN
-if (typeof Number.isNaN === 'undefined') {
-  // eslint-disable-next-line no-self-compare
-  Number.isNaN = value => value !== value;
-}
 
 // To support browsers that do not have Intl.PluralRules (e.g IE11 & Safari 12-),
 // - add npm packaged to package.json: "intl-pluralrules": "^1.3.1",